requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-jazzmin==2.6.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
